requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
psutil
